finished initial version of caching vs. calculating values, including MPI part not yet totally optimised.
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/branches/WolfgangEckhardt@791 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/parallel/DomainDecompDummy.cpp
@@ -98,7 +98,7 @@ void DomainDecompDummy::exchangeMolecules(ParticleContainer* moleculeContainer, 
 		}
 	}
 
-	for (int i =  0; i < moleculesToAdd.size(); i++) {
+	for (size_t i =  0; i < moleculesToAdd.size(); i++) {
 		moleculeContainer->addParticle(moleculesToAdd[i]);
 	}
